+++ all-is-cubes-mesh/src/dynamic/mod.rs
@@ -38,9 +38,7 @@ pub use render_data::{MeshId, RenderDataUpdate};
     clippy::module_name_repetitions,
     reason = "practically needs to be distinct from MeshTypes"
 )]
-pub trait DynamicMeshTypes:
-    MeshTypes<Vertex: Send + Sync + crate::Vertex<Coordinate: Send + Sync>>
-{
+pub trait DynamicMeshTypes: MeshTypes<Vertex: Send + Sync + crate::Vertex> {
     /// Data accompanying meshes within a [`ChunkedSpaceMesh`] and derived from the individual
     /// chunk or block instance meshes, for purposes such as handles to GPU buffers.
     ///